Bug fixes
Fixes:

ability onclick now looks at the correct tag when determining who is friend and foe

ability onclick now targets the correct location

removed the bots cards from needing to be in play

Miles and Spawner builds animator now freezes properly
Cause: 2 animators active at once

Blinder damage from 120 to 50 each

noseDive target position now targets the ground instead of the unit location
Cause: didnt set position.y = 0

units no longer have extended range when retargeting a unit outside its attack range but within its vision range
given its attack is past its attackReadyPercentage and now should lock onto its target given its still within vision range

Mizuichis summon now gets summoned immidiatly, but the summon has summoning sickness instead

Additions:

Units just placed can now have their abilities dragged, but they will be red and cant be cast until its ready

Cards are now selected once you either click them or drag them. Still can only drag them from the tray for now

+++ Project Everdawn/Assets/_Scripts/Managers/GameConstants.cs
@@ -10,7 +10,7 @@ public static class GameConstants
     public static string HUD_CANVAS = "HUD - Canvas";       //The name of the gameObject that holds the UI
     public static string PLAYER_TAG = "Player";
 
-    public static int FLY_ZONE_HEIGHT = 20; //The distance of the fly zone above the ground
+    public static float FLY_ZONE_HEIGHT = 20; //The distance of the fly zone above the ground
     public static float SAMPLE_POSITION_RADIUS = 30f; //The radius used for the Unity function SamplePostion
 
     public enum MOVEMENT_TYPE {
